Refactor sequencer options to make API modules modular
Previously, we had one optional module for the entire API. If you
chose this you would get all the services (query service and submit
API), if you omitted it you would get nothing. This makes it
impossible to run an HTTP server with just a healthcheck endpoint,
for example, or to run just the submit API without the heavyweight
query API.

Now, there is one base module (http) which starts a barebones web
server with healthcheck and version endpoints. There are additional
modules you can add after that one (query and commit) which enable
additional functionality.

Closes #491